refactor(ui): migrate subjects from int32_t to float for precise value representation
Change all UI subject values from integer to floating-point to support
more precise sensor readings and smoother animations.

Changes:
- Update subjects.h function signatures from int32_t to float
- Change analog_dial to use lv_subject_get_float() instead of get_int()
- Convert all subject initializations to lv_subject_init_float()
- Update setter functions to use lv_subject_set_float()
- Add CONFIG_LV_USE_FLOAT=y to sdkconfig.defaults for LVGL float support
- Update mock data task to generate smooth float value transitions
- Update documentation comments to reflect float type requirements

+++ include/ui/subjects.h
@@ -53,31 +53,31 @@ void subjects_deinit(void);
  * @brief Set temperature subject value (thread-safe)
  * @param temperature Temperature value in degrees Celsius
  */
-void subjects_set_temperature(int32_t temperature);
+void subjects_set_temperature(float temperature);
 
 /**
  * @brief Set humidity subject value (thread-safe)
  * @param humidity Humidity value as percentage (0-100)
  */
-void subjects_set_humidity(int32_t humidity);
+void subjects_set_humidity(float humidity);
 
 /**
  * @brief Set heater power subject value (thread-safe)
  * @param power Power level as percentage (0-100)
  */
-void subjects_set_heater_power(int32_t power);
+void subjects_set_heater_power(float power);
 
 /**
  * @brief Set fan speed subject value (thread-safe)
  * @param speed Fan speed as percentage (0-100)
  */
-void subjects_set_fan_speed(int32_t speed);
+void subjects_set_fan_speed(float speed);
 
 /**
  * @brief Set system state subject value (thread-safe)
  * @param state System state (0=idle, 1=heating, 2=cooling, 3=error)
  */
-void subjects_set_system_state(int32_t state);
+void subjects_set_system_state(float state);
 
 #ifdef __cplusplus
 }
